backend/demo-python/requirements.txt
nacos-sdk-python==2.0.1
flask==3.0.2
flask_restx==1.3.0
python-dotenv==1.0.1